requirements.txt
Django==2.1.4
django-ckeditor==5.6.1
django-js-asset==1.1.0
Pillow==5.4.0
psycopg2==2.7.6.1
pytz==2018.7
Unipath==1.1
gunicorn==19.9.0
